requirements.txt
pandas # for dataframe
requests # requests
google-cloud-bigquery # for bigquery 
xlrd # for excel processing